+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - model Activity
 - ActivityDryForm
+- Activities page for guilds
 
 ## [1.1.9] - 2019-12-24
 ### Modified